Move _launch_future logic to LadderGame
