+++ tests/unit_tests/test_game.py
@@ -18,7 +18,6 @@ from server.games import (
     Victory,
     VisibilityState
 )
-from server.games.game import GameClosedError
 from server.games.game_results import ArmyOutcome
 from server.games.typedefs import FeaturedModType
 from server.rating import InclusiveRange, RatingType
@@ -399,21 +398,6 @@ async def test_remove_game_connection(
     game.add_game_connection(mock_game_connection)
     await game.remove_game_connection(mock_game_connection)
     assert players.hosting not in game.players
-    e = game._launch_future.exception()
-    assert isinstance(e, GameClosedError)
-    assert e.player == mock_game_connection.player
-
-
-async def test_do_not_cancel_live_games(
-    game: Game, players, mock_game_connection
-):
-    game.state = GameState.LIVE
-    mock_game_connection.player = players.hosting
-    mock_game_connection.state = GameConnectionState.CONNECTED_TO_HOST
-    game.add_game_connection(mock_game_connection)
-    await game.remove_game_connection(mock_game_connection)
-    assert players.hosting not in game.players
-    assert not game._launch_future.done()
 
 
 async def test_game_end_when_no_more_connections(
